Initial implementation of MFA. It's slow, but it works, and is fairly fail safe.
