+++ app/Events/Robinhood/MultiFactorNecessary.php
@@ -1,0 +1,40 @@
+<?php
+
+namespace App\Events\Robinhood;
+
+use App\User;
+use Illuminate\Queue\SerializesModels;
+use Illuminate\Broadcasting\PrivateChannel;
+use Illuminate\Foundation\Events\Dispatchable;
+use Illuminate\Broadcasting\InteractsWithSockets;
+use Illuminate\Contracts\Broadcasting\ShouldBroadcast;
+
+class MultiFactorNecessary implements ShouldBroadcast
+{
+    use Dispatchable, InteractsWithSockets, SerializesModels;
+
+    private $user;
+    public $message;
+
+    /**
+     * Create a new event instance.
+     *
+     * @param User      $user
+     * @param string    $message
+     */
+    public function __construct(User $user, $message)
+    {
+        $this->user = $user;
+        $this->message = $message;
+    }
+
+    /**
+     * Get the channels the event should broadcast on.
+     *
+     * @return \Illuminate\Broadcasting\Channel|array
+     */
+    public function broadcastOn()
+    {
+        return new PrivateChannel('user.' . $this->user->id);
+    }
+}
