+++ app/MFACode.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace App;
+
+use Illuminate\Database\Eloquent\Model;
+
+class MFACode extends Model
+{
+    protected $table = 'mfa_codes';
+    protected $fillable = ['code', 'user_id'];
+
+    public function user()
+    {
+        return $this->belongsTo(User::class);
+    }
+}
