+++ routes/api.php
@@ -29,6 +29,18 @@ Route::get('stock/{ticker}/price', function (Request $request, $ticker) {
     return $stock->json();
 });
 
+Route::prefix('user')->name('user.')->middleware('auth:api')->group(function () {
+    Route::get('/', function (Request $request) {
+        return $request->user();
+    });
+
+    Route::prefix('{user}')->group(function () {
+        Route::prefix('robinhood')->name('robinhood')->group(function () {
+            Route::post('mfa', 'RobinhoodController@receiveMfaCode');
+        });
+    });
+});
+
 Route::prefix('stocks')->name('stocks.')->middleware('auth:api')->group(function () {
     Route::prefix('{symbol}')->group(function () {
         Route::get('exists', 'StocksController@exists');
